move transaction-related calls from EthService to its own class
