+++ src/main/scala/io/iohk/ethereum/consensus/ethash/EthashBlockCreator.scala
@@ -9,6 +9,7 @@ import io.iohk.ethereum.jsonrpc.AkkaTaskOps.TaskActorOps
 import io.iohk.ethereum.ledger.InMemoryWorldStateProxy
 import io.iohk.ethereum.ommers.OmmersPool
 import io.iohk.ethereum.transactions.PendingTransactionsManager.PendingTransactionsResponse
+import io.iohk.ethereum.transactions.TransactionPicker
 import monix.eval.Task
 import scala.concurrent.duration.FiniteDuration
