+++ src/main/scala/io/iohk/ethereum/nodebuilder/NodeBuilder.scala
@@ -420,6 +420,28 @@ trait EthMiningServiceBuilder {
     txPoolConfig.getTransactionFromPoolTimeout
   )
 }
+trait EthTxServiceBuilder {
+  self: StorageBuilder
+    with BlockchainBuilder
+    // with BlockchainConfigBuilder
+    with PendingTransactionsManagerBuilder
+    with LedgerBuilder
+    // with KeyStoreBuilder
+    // with SyncControllerBuilder
+    // with OmmersPoolBuilder
+    // with ConsensusBuilder
+    // with ConsensusConfigBuilder
+    // with FilterManagerBuilder
+    // with FilterConfigBuilder
+    with TxPoolConfigBuilder =>
+
+  lazy val ethTxService = new EthTxService(
+    blockchain,
+    ledger,
+    pendingTransactionsManager,
+    txPoolConfig.getTransactionFromPoolTimeout
+  )
+}
 
 trait EthBlocksServiceBuilder {
   self: BlockchainBuilder with BlockchainConfigBuilder with LedgerBuilder =>
@@ -511,6 +533,7 @@ trait JSONRpcControllerBuilder {
     with MiningServiceBuilder
     with EthMiningServiceBuilder
     with EthBlocksServiceBuilder
+    with EthTxServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
@@ -530,6 +553,7 @@ trait JSONRpcControllerBuilder {
       ethService,
       ethMiningService,
       ethBlocksService,
+      ethTxService,
       personalService,
       testService,
       debugService,
@@ -711,6 +735,7 @@ trait Node
     with MiningServiceBuilder
     with EthMiningServiceBuilder
     with EthBlocksServiceBuilder
+    with EthTxServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
