+++ src/test/scala/io/iohk/ethereum/jsonrpc/CheckpointingJRCSpec.scala
@@ -210,6 +210,7 @@ class CheckpointingJRCSpec
     val ethService = mock[EthService]
     val ethMiningService = mock[EthMiningService]
     val ethBlocksService = mock[EthBlocksService]
+    val ethTxService = mock[EthTxService]
     val qaService = mock[QAService]
     val checkpointingService = mock[CheckpointingService]
     val mantisService = mock[MantisService]
@@ -221,6 +222,7 @@ class CheckpointingJRCSpec
         ethService,
         ethMiningService,
         ethBlocksService,
+        ethTxService,
         personalService,
         None,
         debugService,
