+++ src/test/scala/io/iohk/ethereum/jsonrpc/EthServiceSpec.scala
@@ -58,134 +58,6 @@ class EthServiceSpec
     Integer.parseInt(protocolVersion.drop(2), 16) shouldEqual currentProtocolVersion
   }
 
-  it should "answer eth_getTransactionByBlockHashAndIndex with None when there is no block with the requested hash" in new TestSetup {
-    val txIndexToRequest = blockToRequest.body.transactionList.size / 2
-    val request = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, txIndexToRequest)
-    val response = ethService.getTransactionByBlockHashAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    response.transactionResponse shouldBe None
-  }
-
-  it should "answer eth_getTransactionByBlockHashAndIndex with None when there is no tx in requested index" in new TestSetup {
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val invalidTxIndex = blockToRequest.body.transactionList.size
-    val requestWithInvalidIndex = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, invalidTxIndex)
-    val response = ethService
-      .getTransactionByBlockHashAndIndex(requestWithInvalidIndex)
-      .runSyncUnsafe(Duration.Inf)
-      .toOption
-      .get
-
-    response.transactionResponse shouldBe None
-  }
-
-  it should "answer eth_getTransactionByBlockHashAndIndex with the transaction response correctly when the requested index has one" in new TestSetup {
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val txIndexToRequest = blockToRequest.body.transactionList.size / 2
-    val request = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, txIndexToRequest)
-    val response = ethService.getTransactionByBlockHashAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    val requestedStx = blockToRequest.body.transactionList.apply(txIndexToRequest)
-    val expectedTxResponse = TransactionResponse(requestedStx, Some(blockToRequest.header), Some(txIndexToRequest))
-    response.transactionResponse shouldBe Some(expectedTxResponse)
-  }
-
-  it should "answer eth_getRawTransactionByBlockHashAndIndex with None when there is no block with the requested hash" in new TestSetup {
-    // given
-    val txIndexToRequest = blockToRequest.body.transactionList.size / 2
-    val request = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, txIndexToRequest)
-
-    // when
-    val response = ethService.getRawTransactionByBlockHashAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    // then
-    response.transactionResponse shouldBe None
-  }
-
-  it should "answer eth_getRawTransactionByBlockHashAndIndex with None when there is no tx in requested index" in new TestSetup {
-    // given
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val invalidTxIndex = blockToRequest.body.transactionList.size
-    val requestWithInvalidIndex = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, invalidTxIndex)
-
-    // when
-    val response = ethService
-      .getRawTransactionByBlockHashAndIndex(requestWithInvalidIndex)
-      .runSyncUnsafe(Duration.Inf)
-      .toOption
-      .value
-
-    // then
-    response.transactionResponse shouldBe None
-  }
-
-  it should "answer eth_getRawTransactionByBlockHashAndIndex with the transaction response correctly when the requested index has one" in new TestSetup {
-    // given
-    blockchain.storeBlock(blockToRequest).commit()
-    val txIndexToRequest = blockToRequest.body.transactionList.size / 2
-    val request = GetTransactionByBlockHashAndIndexRequest(blockToRequest.header.hash, txIndexToRequest)
-
-    // when
-    val response = ethService.getRawTransactionByBlockHashAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    // then
-    val expectedTxResponse = blockToRequest.body.transactionList.lift(txIndexToRequest)
-    response.transactionResponse shouldBe expectedTxResponse
-  }
-
-  it should "handle eth_getRawTransactionByHash if the tx is not on the blockchain and not in the tx pool" in new TestSetup {
-    // given
-    (() => ledger.consensus).expects().returns(consensus)
-    val request = GetTransactionByHashRequest(txToRequestHash)
-
-    // when
-    val response = ethService.getRawTransactionByHash(request).runSyncUnsafe()
-
-    // then
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(Nil))
-
-    response shouldEqual Right(RawTransactionResponse(None))
-  }
-
-  it should "handle eth_getRawTransactionByHash if the tx is still pending" in new TestSetup {
-    // given
-    (() => ledger.consensus).expects().returns(consensus)
-    val request = GetTransactionByHashRequest(txToRequestHash)
-
-    // when
-    val response = ethService.getRawTransactionByHash(request).runToFuture
-
-    // then
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(
-      PendingTransactionsResponse(Seq(PendingTransaction(txToRequestWithSender, System.currentTimeMillis)))
-    )
-
-    response.futureValue shouldEqual Right(RawTransactionResponse(Some(txToRequest)))
-  }
-
-  it should "handle eth_getRawTransactionByHash if the tx was already executed" in new TestSetup {
-    // given
-    (() => ledger.consensus).expects().returns(consensus)
-
-    val blockWithTx = Block(Fixtures.Blocks.Block3125369.header, Fixtures.Blocks.Block3125369.body)
-    blockchain.storeBlock(blockWithTx).commit()
-    val request = GetTransactionByHashRequest(txToRequestHash)
-
-    // when
-    val response = ethService.getRawTransactionByHash(request).runSyncUnsafe()
-
-    // then
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(Nil))
-
-    response shouldEqual Right(RawTransactionResponse(Some(txToRequest)))
-  }
-
   it should "return syncing info if the peer is syncing" in new TestSetup {
     syncingController.setAutoPilot(simpleAutoPilot { case SyncProtocol.GetStatus =>
       SyncProtocol.Status.Syncing(999, Progress(200, 10000), Some(Progress(100, 144)))
@@ -300,98 +172,6 @@ class EthServiceSpec
     response.runSyncUnsafe() shouldEqual Right(GetCodeResponse(ByteString("code code code")))
   }
 
-  it should "return 0 gas price if there are no transactions" in new TestSetup {
-    (appStateStorage.getBestBlockNumber _).expects().returning(42)
-
-    val response = ethService.getGetGasPrice(GetGasPriceRequest())
-    response.runSyncUnsafe() shouldEqual Right(GetGasPriceResponse(0))
-  }
-
-  it should "return average gas price" in new TestSetup {
-    blockchain.saveBestKnownBlocks(42)
-    blockchain
-      .storeBlock(Block(Fixtures.Blocks.Block3125369.header.copy(number = 42), Fixtures.Blocks.Block3125369.body))
-      .commit()
-
-    val response = ethService.getGetGasPrice(GetGasPriceRequest())
-    response.runSyncUnsafe() shouldEqual Right(GetGasPriceResponse(BigInt("20000000000")))
-  }
-
-  it should "getTransactionByBlockNumberAndIndexRequest return transaction by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-    blockchain.saveBestKnownBlocks(blockToRequest.header.number)
-
-    val txIndex: Int = 1
-    val request = GetTransactionByBlockNumberAndIndexRequest(BlockParam.Latest, txIndex)
-    val response = ethService.getTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    val expectedTxResponse =
-      TransactionResponse(blockToRequest.body.transactionList(txIndex), Some(blockToRequest.header), Some(txIndex))
-    response.transactionResponse shouldBe Some(expectedTxResponse)
-  }
-
-  it should "getTransactionByBlockNumberAndIndexRequest return empty response if transaction does not exists when getting by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val txIndex: Int = blockToRequest.body.transactionList.length + 42
-    val request =
-      GetTransactionByBlockNumberAndIndexRequest(BlockParam.WithNumber(blockToRequest.header.number), txIndex)
-    val response = ethService.getTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    response.transactionResponse shouldBe None
-  }
-
-  it should "getTransactionByBlockNumberAndIndex return empty response if block does not exists when getting by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val txIndex: Int = 1
-    val request =
-      GetTransactionByBlockNumberAndIndexRequest(BlockParam.WithNumber(blockToRequest.header.number - 42), txIndex)
-    val response = ethService.getTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    response.transactionResponse shouldBe None
-  }
-
-  it should "getRawTransactionByBlockNumberAndIndex return transaction by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-    blockchain.saveBestKnownBlocks(blockToRequest.header.number)
-
-    val txIndex: Int = 1
-    val request = GetTransactionByBlockNumberAndIndexRequest(BlockParam.Latest, txIndex)
-    val response = ethService.getRawTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    val expectedTxResponse = blockToRequest.body.transactionList.lift(txIndex)
-    response.transactionResponse shouldBe expectedTxResponse
-  }
-
-  it should "getRawTransactionByBlockNumberAndIndex return empty response if transaction does not exists when getting by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val txIndex: Int = blockToRequest.body.transactionList.length + 42
-    val request =
-      GetTransactionByBlockNumberAndIndexRequest(BlockParam.WithNumber(blockToRequest.header.number), txIndex)
-    val response = ethService.getRawTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    response.transactionResponse shouldBe None
-  }
-
-  it should "getRawTransactionByBlockNumberAndIndex return empty response if block does not exists when getting by index" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    blockchain.storeBlock(blockToRequest).commit()
-
-    val txIndex: Int = 1
-    val request =
-      GetTransactionByBlockNumberAndIndexRequest(BlockParam.WithNumber(blockToRequest.header.number - 42), txIndex)
-    val response = ethService.getRawTransactionByBlockNumberAndIndex(request).runSyncUnsafe(Duration.Inf).toOption.get
-
-    response.transactionResponse shouldBe None
-  }
-
   it should "handle getBalance request" in new TestSetup {
     (() => ledger.consensus).expects().returns(consensus)
     val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
@@ -488,127 +268,6 @@ class EthServiceSpec
     val response = ethService.getTransactionCount(GetTransactionCountRequest(address, BlockParam.Latest))
 
     response.runSyncUnsafe() shouldEqual Right(GetTransactionCountResponse(BigInt(999)))
-  }
-
-  it should "handle get transaction by hash if the tx is not on the blockchain and not in the tx pool" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-
-    val request = GetTransactionByHashRequest(txToRequestHash)
-    val response = ethService.getTransactionByHash(request).runSyncUnsafe()
-
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(Nil))
-
-    response shouldEqual Right(GetTransactionByHashResponse(None))
-  }
-
-  it should "handle get transaction by hash if the tx is still pending" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-
-    val request = GetTransactionByHashRequest(txToRequestHash)
-    val response = ethService.getTransactionByHash(request).runToFuture
-
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(
-      PendingTransactionsResponse(Seq(PendingTransaction(txToRequestWithSender, System.currentTimeMillis)))
-    )
-
-    response.futureValue shouldEqual Right(GetTransactionByHashResponse(Some(TransactionResponse(txToRequest))))
-  }
-
-  it should "handle get transaction by hash if the tx was already executed" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-
-    val blockWithTx = Block(Fixtures.Blocks.Block3125369.header, Fixtures.Blocks.Block3125369.body)
-    blockchain.storeBlock(blockWithTx).commit()
-
-    val request = GetTransactionByHashRequest(txToRequestHash)
-    val response = ethService.getTransactionByHash(request).runSyncUnsafe()
-
-    pendingTransactionsManager.expectMsg(PendingTransactionsManager.GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(Nil))
-
-    response shouldEqual Right(
-      GetTransactionByHashResponse(Some(TransactionResponse(txToRequest, Some(blockWithTx.header), Some(0))))
-    )
-  }
-
-  it should "calculate correct contract address for contract creating by transaction" in new TestSetup {
-    val body = BlockBody(Seq(Fixtures.Blocks.Block3125369.body.transactionList.head, contractCreatingTransaction), Nil)
-    val blockWithTx = Block(Fixtures.Blocks.Block3125369.header, body)
-    val gasUsedByTx = 4242
-    blockchain
-      .storeBlock(blockWithTx)
-      .and(
-        blockchain.storeReceipts(
-          Fixtures.Blocks.Block3125369.header.hash,
-          Seq(fakeReceipt, fakeReceipt.copy(cumulativeGasUsed = fakeReceipt.cumulativeGasUsed + gasUsedByTx))
-        )
-      )
-      .commit()
-
-    val request = GetTransactionReceiptRequest(contractCreatingTransaction.hash)
-    val response = ethService.getTransactionReceipt(request)
-
-    response.runSyncUnsafe() shouldEqual Right(
-      GetTransactionReceiptResponse(
-        Some(
-          TransactionReceiptResponse(
-            receipt = fakeReceipt.copy(cumulativeGasUsed = fakeReceipt.cumulativeGasUsed + gasUsedByTx),
-            stx = contractCreatingTransaction,
-            signedTransactionSender = contractCreatingTransactionSender,
-            transactionIndex = 1,
-            blockHeader = Fixtures.Blocks.Block3125369.header,
-            gasUsedByTransaction = gasUsedByTx
-          )
-        )
-      )
-    )
-  }
-
-  it should "send message to pendingTransactionsManager and return an empty GetPendingTransactionsResponse" in new TestSetup {
-    val res = ethService.getTransactionsFromPool.runSyncUnsafe()
-
-    pendingTransactionsManager.expectMsg(GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(Nil))
-
-    res shouldBe PendingTransactionsResponse(Nil)
-  }
-
-  it should "send message to pendingTransactionsManager and return GetPendingTransactionsResponse with two transactions" in new TestSetup {
-    val transactions = (0 to 1)
-      .map(_ => {
-        val fakeTransaction = SignedTransactionWithSender(
-          Transaction(
-            nonce = 0,
-            gasPrice = 123,
-            gasLimit = 123,
-            receivingAddress = Address("0x1234"),
-            value = 0,
-            payload = ByteString()
-          ),
-          signature = ECDSASignature(0, 0, 0.toByte),
-          sender = Address("0x1234")
-        )
-        PendingTransaction(fakeTransaction, System.currentTimeMillis)
-      })
-      .toList
-
-    val res = ethService.getTransactionsFromPool.runToFuture
-
-    pendingTransactionsManager.expectMsg(GetPendingTransactions)
-    pendingTransactionsManager.reply(PendingTransactionsResponse(transactions))
-
-    res.futureValue shouldBe PendingTransactionsResponse(transactions)
-  }
-
-  it should "send message to pendingTransactionsManager and return an empty GetPendingTransactionsResponse in case of error" in new TestSetup {
-    val res = ethService.getTransactionsFromPool.runSyncUnsafe()
-
-    pendingTransactionsManager.expectMsg(GetPendingTransactions)
-    pendingTransactionsManager.reply(new ClassCastException("error"))
-
-    res shouldBe PendingTransactionsResponse(Nil)
   }
 
   // NOTE TestSetup uses Ethash consensus; check `consensusConfig`.
