+++ src/test/scala/io/iohk/ethereum/jsonrpc/JsonRpcControllerFixture.scala
@@ -101,8 +101,31 @@ class JsonRpcControllerFixture(implicit system: ActorSystem)
     pendingTransactionsManager.ref,
     getTransactionFromPoolTimeout
   )
+  val ethTxService = new EthTxService(
+    blockchain,
+    ledger,
+    pendingTransactionsManager.ref,
+    getTransactionFromPoolTimeout
+  )
 
   val ethBlocksService = new EthBlocksService(blockchain, ledger, blockchainConfig)
+
+  protected def newJsonRpcController(ethBlocksService: EthBlocksService) =
+    new JsonRpcController(
+      web3Service,
+      netService,
+      ethService,
+      ethMiningService,
+      ethBlocksService,
+      ethTxService,
+      personalService,
+      None,
+      debugService,
+      qaService,
+      checkpointingService,
+      mantisService,
+      config
+    )
 
   protected def newJsonRpcController(ethService: EthService, proofService: ProofService = ProofServiceDummy) =
     new JsonRpcController(
@@ -111,6 +134,24 @@ class JsonRpcControllerFixture(implicit system: ActorSystem)
       ethService,
       ethMiningService,
       ethBlocksService,
+      ethTxService,
+      personalService,
+      None,
+      debugService,
+      qaService,
+      checkpointingService,
+      mantisService,
+      config
+    )
+
+  protected def newJsonRpcController(ethTxService: EthTxService) =
+    new JsonRpcController(
+      web3Service,
+      netService,
+      ethService,
+      ethMiningService,
+      ethBlocksService,
+      ethTxService,
       personalService,
       None,
       debugService,
@@ -122,7 +163,7 @@ class JsonRpcControllerFixture(implicit system: ActorSystem)
       config
     )
 
-  val jsonRpcController = newJsonRpcController(ethService, ethBlocksService)
+  val jsonRpcController = newJsonRpcController(ethService)
 
   val blockHeader = Fixtures.Blocks.ValidBlock.header.copy(
     logsBloom = BloomFilter.EmptyBloomFilter,
